% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biaa.R
\name{biaa}
\alias{biaa}
\title{Compute archetypes from multivariate data}
\usage{
biaa(
  data,
  k,
  c,
  nrep = 10,
  maxiter = 10000,
  minimpr = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{data}{A \eqn{n \times m} matrix where \eqn{n} is the number of observations and \eqn{m}
is the number of variables.}

\item{k}{A list with the desired numbers of archertypes.}

\item{c}{A list with the desired numbers of variables.}

\item{nrep}{The number of repetitions that each scenario is computed.}

\item{maxiter}{The maximum number of iterations.}

\item{minimpr}{The minimal value of improvement between two iterations.}
}
\value{
A matrix with all possible scenarios (determined by the Cartesian product between
\code{k} and\code{ c}).
}
\description{
Perform bi-archetypal analysis on a multivariate dataset in order to obtain the archetypes.
}
\examples{
biaa(data, 4, 10)
biaa(data, 1:6, 1:15)
biaa(data, 1:5, 5:8, nrep=100)
biaa(data, 4:8, 10, maxiter=1000, minimpr=0.0001)

}
