% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biaa.R
\name{getmodel}
\alias{getmodel}
\title{Get the archetypal model}
\usage{
getmodel(data, k, c)
}
\arguments{
\item{data}{A \code{"biaaGroup"} class list.}

\item{k}{The number of archetypes.}

\item{c}{The number of variables.}
}
\description{
Get a specific archetypal model stored in a \code{"biaaGroup"} class list.
}
\examples{
getmodel(archetypes, 3, 5)

}
