% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss_biaa.R
\name{rssplot}
\alias{rssplot}
\title{Plot the bi-archetypal analysis RSS}
\usage{
rssplot(data, phi = 30, theta = 135)
}
\arguments{
\item{data}{A \code{"biaaGroup"} class list.}

\item{phi}{A number giving the azimuthal direction.}

\item{theta}{A number giving the colatitude.}
}
\description{
Plot in a the RSS of the all archetypes stored in a \code{"biaaGroup"} class list.
}
\examples{
rssplot(archetypes)
rssplot(archetypes, phi = 45, theta = 165)

}
